refactor: medicine 패키지를 medication으로 rename
